+++ addons/dialogic/Editor/ThemeEditor/ThemeEditor.gd
@@ -88,6 +88,7 @@ onready var n : Dictionary = {
 	'name_shadow_offset_x': $"VBoxContainer/TabContainer/Name Label/Column/GridContainer/HBoxContainer/ShadowOffsetX",
 	'name_shadow_offset_y': $"VBoxContainer/TabContainer/Name Label/Column/GridContainer/HBoxContainer/ShadowOffsetY",
 	'name_bottom_gap': $"VBoxContainer/TabContainer/Name Label/Column3/GridContainer/HBoxContainer5/BottomGap",
+	'name_horizontal_offset': $"VBoxContainer/TabContainer/Name Label/Column3/GridContainer/HBoxContainer5/HorizontalOffset",
 	'name_background_modulation': $"VBoxContainer/TabContainer/Name Label/Column2/GridContainer/HBoxContainer6/CheckBox",
 	'name_background_modulation_color': $"VBoxContainer/TabContainer/Name Label/Column2/GridContainer/HBoxContainer6/ColorPickerButton",
 	'name_padding_x': $"VBoxContainer/TabContainer/Name Label/Column2/GridContainer/HBoxContainer/NamePaddingX",
@@ -312,6 +313,7 @@ func load_theme(filename):
 	n['name_shadow_offset_x'].value = theme.get_value('name', 'shadow_offset', Vector2(2,2)).x
 	n['name_shadow_offset_y'].value = theme.get_value('name', 'shadow_offset', Vector2(2,2)).y
 	n['name_bottom_gap'].value = theme.get_value('name', 'bottom_gap', 48)
+	n['name_horizontal_offset'].value = theme.get_value('name', 'horizontal_offset', 0)
 	
 	n['name_position'].select(theme.get_value('name', 'position', 0))
 	
@@ -761,8 +763,9 @@ func _on_name_padding_value_changed(_value) -> void:
 
 func _on_name_BottomGap_value_changed(value) -> void:
 	if loading:
-		return
-	DialogicResources.set_theme_value(current_theme, 'name', 'bottom_gap', value)
+		return	
+	DialogicResources.set_theme_value(current_theme, 'name', 'bottom_gap', n['name_bottom_gap'].value)
+	DialogicResources.set_theme_value(current_theme, 'name', 'horizontal_offset', n['name_horizontal_offset'].value)
 	_on_PreviewButton_pressed() # Refreshing the preview