+++ addons/dialogic/Nodes/TextBubble.gd
@@ -187,12 +187,15 @@ func start_text_timer():
 
 func align_name_label():
 	var name_padding = _theme.get_value('name', 'name_padding', Vector2( 10, 0 ))
+	var horizontal_offset = _theme.get_value('name', 'horizontal_offset', 0)
 	var name_label_position = _theme.get_value('name', 'position', 0)
 	var label_size = name_label.rect_size.x
-	if name_label_position == 1: # Center
-		name_label.rect_global_position.x = rect_global_position.x + (rect_size.x / 2) - (label_size / 2)
+	if name_label_position == 0:
+		name_label.rect_global_position.x = rect_global_position.x + horizontal_offset
+	elif name_label_position == 1: # Center
+		name_label.rect_global_position.x = rect_global_position.x + (rect_size.x / 2) - (label_size / 2) + horizontal_offset
 	elif name_label_position == 2: # Right
-		name_label.rect_global_position.x = rect_global_position.x + rect_size.x - label_size
+		name_label.rect_global_position.x = rect_global_position.x + rect_size.x - label_size + horizontal_offset
 
 ## *****************************************************************************
 ##								OVERRIDES