+++ docs/changelog.md
@@ -15,6 +15,7 @@
   - Added a new option to make the dialog backgrounds full width
   - You can now set a character for the preview message
   - Three positions for the name label: Left, Center and Right
+  - You can set the verticall and horizontal offset of the name label
 - Character Editor
   - You can now add a scene as a portrait, so you can now use AnimatedSprite or whatever you might need
   - You can set nicknames to characters [[zakary93](https://github.com/zakary93)]